--- Horologion/src/utils_time.cpp
@@ -1,26 +1,0 @@
-#include "utils_time.hpp"
-
-#include "logger.hpp"
-
-std::string epoch_time_to_ascii_time(const std::time_t &epoch_time)
-{
-    std::string ascii_time = std::string(std::asctime(std::localtime(&epoch_time)));
-
-    if (not ascii_time.empty() and ascii_time[ascii_time.length() - 1] == '\n') {
-        ascii_time.erase(ascii_time.length() - 1);
-    }
-
-    return ascii_time;
-}
-
-std::time_t get_epoch_time_from_configs(const int &hour, const int &minute, const int &second)
-{
-    std::time_t now = std::time(nullptr);
-    std::tm *tm_time = std::localtime(&now);
-
-    tm_time->tm_hour = hour;
-    tm_time->tm_min = minute;
-    tm_time->tm_sec = second;
-
-    return mktime(tm_time);
-}
