--- Horologion/src/utils_time.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <ctime>
-#include <string>
-
-std::string epoch_time_to_ascii_time(const std::time_t &epoch_time);
-std::time_t get_epoch_time_from_configs(const int &hour, const int &minute, const int &second);
